// frontend/src/api/model/integrationStatus.ts
/**
 * Efficient Labeling of Time Series
 * No description provided
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An enumeration.
 */
export type IntegrationStatus = 'integrating' | 'error' | 'finished';

export const IntegrationStatus = {
    Integrating: 'integrating' as IntegrationStatus,
    Error: 'error' as IntegrationStatus,
    Finished: 'finished' as IntegrationStatus
};